requirements.txt
asgiref==3.4.1
certifi==2022.9.24
charset-normalizer==2.0.12
Django==3.2.16
django-appconf==1.0.5
django-cors-headers==3.10.1
django-ipware==4.0.2
djangorestframework==3.14.0
djangorestframework-recaptcha==0.2.0
drf-recaptcha==2.0.7
idna==3.4
pytz==2022.6
requests==2.27.1
sqlparse==0.4.3
typing_extensions==4.1.1
urllib3==1.26.12
